Added Export as XYZ Functionality
Can now export as an XYZ file
Also added a few checks in the import/export method to ensure that there
is in fact a file selected to import from/export to. Definitely needs
some cleaning up.
Also switched importer/exporter to use File references instead of the
absolute path as a string